# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /admin/audit-logs:
    get:
      operationId: AdminController_auditLogs
      parameters:
      - name: actorId
        required: false
        in: query
        schema:
          type: string
      - name: action
        required: false
        in: query
        schema:
          type: string
      - name: resourceType
        required: false
        in: query
        schema:
          type: string
      - name: from
        required: false
        in: query
        schema:
          type: string
      - name: to
        required: false
        in: query
        schema:
          type: string
      - name: page
        required: false
        in: query
        schema:
          minimum: 1
          type: number
      - name: limit
        required: false
        in: query
        schema:
          minimum: 1
          maximum: 200
          type: number
      responses:
        '200':
          description: ''
      summary: Query audit logs
      tags:
      - Admin
...
  /admin/bookings/{id}/generate-ticket-pdf:
    post:
      operationId: AdminController_generateBookingPdf
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '201':
          description: ''
      summary: Enqueue ticket PDF generation for a booking
      tags:
      - Admin
...
  /admin/organizer-requests:
    get:
      operationId: AdminController_listOrganizerRequests
      parameters:
      - name: query
        required: false
        in: query
        schema:
          type: string
      - name: page
        required: false
        in: query
        schema:
          minimum: 1
          type: number
      - name: limit
        required: false
        in: query
        schema:
          minimum: 1
          maximum: 200
          type: number
      - name: isSuspended
        required: false
        in: query
        schema:
          type: boolean
      - name: organizerStatus
        required: false
        in: query
        schema:
          type: string
      responses:
        '200':
          description: ''
      summary: List organizer requests
      tags:
      - Admin
  /admin/organizer-requests/{id}:
    patch:
      operationId: AdminController_decideOrganizerRequest
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizerRequestDecisionDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizerApplication'
      summary: Approve or reject an organizer request
      tags:
      - Admin
...
  /admin/platform-settings:
    get:
      operationId: AdminController_getPlatformSettings
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
      summary: Get platform settings
      tags:
      - Admin
    patch:
      operationId: AdminController_updatePlatformSettings
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePlatformSettingsDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
      summary: Update platform settings
      tags:
      - Admin
...
  /admin/referrals:
    get:
      operationId: AdminController_listReferrals
      parameters:
      - name: status
        required: false
        in: query
        schema:
          type: string
      - name: referrerId
        required: false
        in: query
        schema:
          type: string
      - name: startDate
        required: false
        in: query
        schema:
          type: string
      - name: endDate
        required: false
        in: query
        schema:
          type: string
      - name: query
        required: false
        in: query
        schema:
          type: string
      - name: page
        required: false
        in: query
        schema:
          minimum: 1
          type: number
      - name: limit
        required: false
        in: query
        schema:
          minimum: 1
          maximum: 200
          type: number
      responses:
        '200':
          description: ''
      summary: List all referrals with filters
      tags:
      - Admin
  /admin/referrals/codes:
    get:
      operationId: AdminController_listReferralCodes
      parameters:
      - name: query
        required: false
        in: query
        schema:
          type: string
      - name: tier
        required: false
        in: query
        schema:
          type: string
      - name: page
        required: false
        in: query
        schema:
          minimum: 1
          type: number
      - name: limit
        required: false
        in: query
        schema:
          minimum: 1
          maximum: 200
          type: number
      responses:
        '200':
          description: ''
      summary: List all referral codes with stats
      tags:
      - Admin
...
  /admin/treks/pending:
    get:
      operationId: AdminController_listPendingTreks
      parameters:
      - name: query
        required: false
        in: query
        schema:
          type: string
      - name: page
        required: false
        in: query
        schema:
          minimum: 1
          type: number
      - name: limit
        required: false
        in: query
        schema:
          minimum: 1
          maximum: 200
          type: number
      - name: isSuspended
        required: false
        in: query
        schema:
          type: boolean
      - name: organizerStatus
        required: false
        in: query
        schema:
          type: string
      responses:
        '200':
          description: ''
      summary: List pending treks for approval
      tags:
      - Admin
  /admin/treks/{id}/decision:
    patch:
      operationId: AdminController_decideTrek
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrekDecisionDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trek'
      summary: Approve or reject a trek
      tags:
      - Admin
...
  /admin/users:
    get:
      operationId: AdminController_listUsers
      parameters:
      - name: query
        required: false
        in: query
        schema:
          type: string
      - name: page
        required: false
        in: query
        schema:
          minimum: 1
          type: number
      - name: limit
        required: false
        in: query
        schema:
          minimum: 1
          maximum: 200
          type: number
      - name: isSuspended
        required: false
        in: query
        schema:
          type: boolean
      - name: organizerStatus
        required: false
        in: query
        schema:
          type: string
      responses:
        '200':
          description: ''
      summary: List users
      tags:
      - Admin
...
  /admin/users/{id}/status:
    patch:
      operationId: AdminController_updateUserStatus
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserStatusDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      summary: Update user status
      tags:
      - Admin
...
  /friend-requests:
    post:
      operationId: FriendshipsController_sendRequest
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendFriendRequestDto'
      responses:
        '200':
          description: Friend request sent
      summary: Send a friend request
      tags:
      - Friendships
    get:
      operationId: FriendshipsController_getPendingRequests
      parameters: []
      responses:
        '200':
          description: List of pending requests
      summary: List incoming pending friend requests
      tags:
      - Friendships
...
  /friend-requests/{id}/accept:
    post:
      operationId: FriendshipsController_acceptRequest
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: Friend request accepted
      summary: Accept a friend request
      tags:
      - Friendships
  /friend-requests/{id}/decline:
    post:
      operationId: FriendshipsController_declineRequest
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: Friend request declined
      summary: Decline a friend request
      tags:
      - Friendships
...
  /friends/{id}:
    delete:
      operationId: FriendshipsController_removeFriend
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: Friend removed
      summary: Remove a friend
      tags:
      - Friendships
...
info:
  title: Offbeat प्रवासी API
  description: Full-scale, production-ready trek discovery and booking platform.
  version: 1.0.0
  contact: {}
...
servers:
- url: http://localhost:4000
  description: Local development
- url: https://api.offbeatpravasi.com
  description: Production
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
    ApiKey:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
...
    DeviceToken:
      type: object
      properties:
        id:
          type: string
          default: fd9aba86-18cd-437a-a382-120b4b0a33e3
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
        platform:
          type: string
          enum:
          - ANDROID
          - IOS
          - WEB
        appVersion:
          type: string
        lastSeenAt:
          format: date-time
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
      - id
      - user
      - token
      - platform
      - createdAt
      - updatedAt
...
    GearItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          enum:
          - CLOTHING
          - FOOTWEAR
          - CAMPING
          - NAVIGATION
          - TOILETRIES
          - DOCUMENTS
          - OPTIONAL
          type: string
        isActive:
          type: boolean
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
      - id
      - name
      - category
      - isActive
      - createdAt
      - updatedAt
...
    ItineraryDay:
      type: object
      properties:
        id:
          type: string
        trekId:
          type: string
        trek:
          $ref: '#/components/schemas/Trek'
        dayNumber:
          type: number
        title:
          type: string
        description:
          type: string
          nullable: true
        distanceKm:
          type: number
          nullable: true
        altitudeGainM:
          type: number
          nullable: true
        altitudeLossM:
          type: number
          nullable: true
        maxAltitudeM:
          type: number
          nullable: true
        mealPlan:
          type: object
          properties:
            breakfast:
              type: string
            lunch:
              type: string
            dinner:
              type: string
          required: []
        accommodationType:
          type: string
          nullable: true
          enum:
          - TENT
          - GUESTHOUSE
          - CAMP
          - HOMESTAY
        activityType:
          type: string
          enum:
          - TREKKING
          - REST
          - ACCLIMATIZATION
          - SIGHTSEEING
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
      - id
      - trekId
      - trek
      - dayNumber
      - title
      - description
      - distanceKm
      - altitudeGainM
      - altitudeLossM
      - maxAltitudeM
      - mealPlan
      - accommodationType
      - activityType
      - createdAt
      - updatedAt
...
    Notification:
      type: object
      properties:
        id:
          type: string
          default: 774b6963-e4ea-4022-a5ee-0a78f4c6d50f
        user:
          $ref: '#/components/schemas/User'
        type:
          type: string
          enum:
          - BOOKING_CONFIRMED
          - BOOKING_CANCELLED
          - BOOKING_REMINDER
          - FRIEND_REQUEST_RECEIVED
          - FRIEND_REQUEST_ACCEPTED
          - NEW_FOLLOWER
          - NEW_LIKE
          - NEW_COMMENT
          - TREK_UPDATE
          - ORGANIZER_APPROVED
          - ORGANIZER_REJECTED
          - STORY_EXPIRING
          - CHECK_IN_REMINDER
          - SAFETY_ALERT
          - EMERGENCY_ESCALATION
          - GROUP_INVITE
          - GROUP_UPDATE
          - WISHLIST_PRICE_DROP
          - BROADCAST
        title:
          type: string
        body:
          type: string
        data:
          type: object
          nullable: true
        readAt:
          format: date-time
          type: string
        createdAt:
          format: date-time
          type: string
      required:
      - id
      - user
      - type
      - title
      - body
      - createdAt
...
    OrganizerApplication:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        user:
          $ref: '#/components/schemas/User'
        organizationName:
          type: string
        contactPerson:
          type: string
        contactPhone:
          type: string
        website:
          type: string
        bio:
          type: string
        yearsOfExperience:
          type: number
        documentUrls:
          type: object
        certificateUrls:
          type: object
        status:
          type: string
          enum:
          - NONE
          - PENDING
          - APPROVED
          - REJECTED
        adminNotes:
          type: string
        submittedAt:
          format: date-time
          type: string
        reviewedAt:
          format: date-time
          type: string
      required:
      - id
      - userId
      - user
      - organizationName
      - contactPhone
      - bio
      - status
      - submittedAt
...
    OrganizerRequestDecisionDto:
      type: object
      properties:
        decision:
          type: string
          enum:
          - APPROVE
          - REJECT
        note:
          type: string
      required:
      - decision
...
    SendFriendRequestDto:
      type: object
      properties:
        receiverId:
          type: string
          format: uuid
      required:
      - receiverId
...
    Trek:
      type: object
      properties:
        id:
          type: string
        organizer:
          nullable: true
          type: object
          allOf:
          - $ref: '#/components/schemas/User'
        name:
          type: string
        slug:
          type: string
          nullable: true
        shortDescription:
          type: string
          nullable: true
        fullDescription:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        geom:
          type: string
          nullable: true
        startDate:
          format: date-time
          type: string
          nullable: true
        endDate:
          format: date-time
          type: string
          nullable: true
        difficulty:
          type: string
          nullable: true
          enum:
          - EASY
          - MODERATE
          - DIFFICULT
          - EXTREME
        costInr:
          type: number
        maxParticipants:
          type: number
        isPublished:
          type: boolean
        currentParticipants:
          type: number
        status:
          type: string
          enum:
          - DRAFT
          - PUBLISHED
          - CANCELLED
          - COMPLETED
        avgRating:
          type: number
        ratingCount:
          type: number
        popularityScore:
          type: number
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TrekTag'
        images:
          type: array
          items:
            $ref: '#/components/schemas/TrekImage'
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/TrekReview'
        itineraryDays:
          type: array
          items:
            $ref: '#/components/schemas/ItineraryDay'
        gearItems:
          type: array
          items:
            $ref: '#/components/schemas/TrekGearItem'
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        deletedAt:
          format: date-time
          type: string
          nullable: true
      required:
      - id
      - organizer
      - name
      - slug
      - shortDescription
      - fullDescription
      - state
      - location
      - latitude
      - longitude
      - geom
      - startDate
      - endDate
      - difficulty
      - costInr
      - maxParticipants
      - isPublished
      - currentParticipants
      - status
      - avgRating
      - ratingCount
      - popularityScore
      - tags
      - images
      - reviews
      - itineraryDays
      - gearItems
      - createdAt
      - updatedAt
      - deletedAt
...
    TrekDecisionDto:
      type: object
      properties:
        decision:
          type: string
          enum:
          - APPROVE
          - REJECT
        note:
          type: string
      required:
      - decision
    TrekGearItem:
      type: object
      properties:
        id:
          type: string
        trekId:
          type: string
        trek:
          $ref: '#/components/schemas/Trek'
        gearItemId:
          type: string
        gearItem:
          $ref: '#/components/schemas/GearItem'
        requirementType:
          enum:
          - REQUIRED
          - RECOMMENDED
          - PROVIDED
          - RENTAL
          type: string
        rentalPriceInr:
          type: number
        notes:
          type: string
        sortOrder:
          type: number
      required:
      - id
      - trekId
      - trek
      - gearItemId
      - gearItem
      - requirementType
      - sortOrder
...
    TrekImage:
      type: object
      properties:
        id:
          type: string
        trek:
          $ref: '#/components/schemas/Trek'
        key:
          type: string
        url:
          type: string
          nullable: true
        isPrimary:
          type: boolean
        order:
          type: number
        altText:
          type: string
          nullable: true
        createdAt:
          format: date-time
          type: string
      required:
      - id
      - trek
      - key
      - url
      - isPrimary
      - order
      - altText
      - createdAt
    TrekReview:
      type: object
      properties:
        id:
          type: string
        trek:
          $ref: '#/components/schemas/Trek'
        user:
          $ref: '#/components/schemas/User'
        rating:
          type: number
        comment:
          type: string
          nullable: true
        createdAt:
          format: date-time
          type: string
      required:
      - id
      - trek
      - user
      - rating
      - comment
      - createdAt
    TrekTag:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        usageCount:
          type: number
        treks:
          type: array
          items:
            $ref: '#/components/schemas/Trek'
      required:
      - id
      - name
      - usageCount
      - treks
...
    UpdatePlatformSettingsDto:
      type: object
      properties:
        settings:
          type: object
        message:
          type: string
      required:
      - settings
...
    UpdateUserStatusDto:
      type: object
      properties:
        isSuspended:
          type: boolean
        organizerStatus:
          type: string
        note:
          type: string
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        passwordHash:
          type: string
          nullable: true
        fullName:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
          enum:
          - MALE
          - FEMALE
          - OTHER
        dateOfBirth:
          format: date-time
          type: string
          nullable: true
        profileImageUrl:
          type: string
          nullable: true
        bannerImageUrl:
          type: string
          nullable: true
        isAdmin:
          type: boolean
        role:
          type: string
          nullable: true
          enum:
          - superadmin
          - moderator
          - finance
          - support
          - analyst
        organizerStatus:
          type: string
          enum:
          - NONE
          - PENDING
          - APPROVED
          - REJECTED
        emailVerified:
          type: boolean
        emailVerifiedAt:
          format: date-time
          type: string
          nullable: true
        userPoints:
          type: number
        userDistanceTravelled:
          type: number
        organizerRating:
          type: number
        isOrganizerActive:
          type: boolean
        isSuspended:
          type: boolean
        deviceTokens:
          type: array
          items:
            $ref: '#/components/schemas/DeviceToken'
        notifications:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
      - id
      - email
      - passwordHash
      - fullName
      - username
      - phone
      - location
      - gender
      - dateOfBirth
      - profileImageUrl
      - bannerImageUrl
      - isAdmin
      - role
      - organizerStatus
      - emailVerified
      - emailVerifiedAt
      - userPoints
      - userDistanceTravelled
      - organizerRating
      - isOrganizerActive
      - isSuspended
      - createdAt
      - updatedAt
...
security:
- ApiKey: []